Refactor: Move collision logic to Vehicle.Collision.cpp
Moves collision detection and reverser car functions to a dedicated file for it
