+++ src/openrct2/libopenrct2.vcxproj
@@ -1127,6 +1127,7 @@
     <ClCompile Include="ride\TrackStyle.cpp" />
     <ClCompile Include="ride\TrainManager.cpp" />
     <ClCompile Include="ride\Vehicle.BoatHire.cpp" />
+    <ClCompile Include="ride\Vehicle.Collision.cpp" />
     <ClCompile Include="ride\Vehicle.Crash.cpp" />
     <ClCompile Include="ride\Vehicle.cpp" />
     <ClCompile Include="ride\Vehicle.Animation.cpp" />
